+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # LDAPClient Changelog
 
+[2021-09-02] Version 1.4
+   - Replaced search's fullDNs parameter with resolveDNs
+   - New resolveDN method that will resolve a DN to an id
+   - Overhauled the member method to allow matching in both directions
+
 [2021-08-28] Version 1.3
    - Initial public release